example-basic/requirements.txt
-i https://abetlen.github.io/llama-cpp-python/whl/cpu
llama-cpp-python